demo_projects/databricks-ai-newsletter/requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dateutil>=2.8.2
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
